test-twitter/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
cffi==1.15.1
cryptography==38.0.4
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
filelock==3.8.0
gunicorn==20.1.0
pbr==5.11.0
pipenv==2022.11.11
platformdirs==2.5.4
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
stevedore==4.1.1
tomli==2.0.1
virtualenv==20.16.7
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
whitenoise==6.2.0
